Move fake slot manager to its own file.
